backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
google-generativeai==0.3.1
pandas==2.1.0
Werkzeug==2.3.7
pathlib==1.0.1
Pillow==10.0.0
python-dotenv==1.0.0